+++ src/templates/index.ts
@@ -1,0 +1,4 @@
+export * from './mdc-configs'
+export * from './mdc-highlighter'
+export * from './mdc-imports'
+export * from './mdc-shiki-bundle'
